requirements.txt
certifi==2020.12.5
click==7.1.2
fastapi==0.63.0
h11==0.11.0
kafka-python==2.0.2
pydantic==1.10.7
starlette==0.13.6
typing-extensions==4.5.0
uvicorn==0.13.3
